Safari-611.1.10/LayoutTests/platform/ipad/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-non-integer-innerwidth-expected.txt


PASS features "innerwidth=/404" should NOT set "width=404"
PASS features "innerwidth=_404" should NOT set "width=404"
PASS features "innerwidth=L404" should NOT set "width=404"
FAIL features "innerwidth=405.5" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405.5 value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405.32" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405.32 value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405LLl" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405LLl value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405^4" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405^4 value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405*3" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405*3 value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405/5" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405/5 value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405  " should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405   value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405e1" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405e1 value after first non-digit will be ignored" expected 405 but got 800
FAIL features "innerwidth=405e-1" should set "width=405" assert_equals: "top=0,left=0,height=401,innerwidth=405e-1 value after first non-digit will be ignored" expected 405 but got 800
